Allow merging stacks with unknown but identical known and bknown
This is the original version of
<https://github.com/Chris-plus-alphanumericgibberish/dNAO/pull/2258>.
A watered down version that only affects rknown and dknown was merged
into dNetHack.

This is a very good QoL change so you don't keep needing to return to
an altar to make your items stack.  It also fixes an info
leak/exploit/annoying optimal strategy of keeping a single item in a
stack un-BUC-tested to identify other items by making it a less
annoying to use feature.

NetHack 3.7, xNetHack, and NetHack 4 all have this change.